backend/ai_backend/agents/rag_agent/requirements.txt
# RAG Agent Dependencies
# Install with: pip install -r requirements.txt

# LangChain core and community packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Vector store
chromadb>=0.4.22

# PDF processing
pypdf>=4.0.0

# OpenAI for embeddings
openai>=1.10.0

# Gemini for condition extraction
google-generativeai>=0.3.0

# Environment variables
python-dotenv>=1.0.0

# Additional utilities
tiktoken>=0.5.2

# Parallel processing with Ray
ray>=2.9.0
